Moving getServerHostName() to KernelUtils class
- Splitting up chained StringBuilder
